backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database drivers
motor==3.3.2
aiomysql==0.2.0
pymongo==4.6.0

# AI and ML
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6

# Data processing and visualization
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
